Update al quiz, el quiz ya quedo integrado con la logica de la bd al igual que el ranking, queda faltando organizar el css de Quiz.jsx, Ranking.jsx y tambien de Quiz3D.jsx
